requirements.txt
torch==2.0.0
torchvision==0.15.1
gradio==3.39.0
pillow==9.5.0
numpy==1.23.5